DSND-2433: Refactor pre-transform mapping and child transformation logic
* All pre-transform mapping is done before transform, including children.
* Transformation of children is handled by checking if each of the child arrays exist.
