+++ src/main/java/uk/gov/companieshouse/filinghistory/consumer/service/FilingHistoryDeltaProcessor.java
@@ -1,32 +1,30 @@
 package uk.gov.companieshouse.filinghistory.consumer.service;
 
-import static uk.gov.companieshouse.api.filinghistory.InternalData.TransactionKindEnum.TOP_LEVEL;
-
+import com.fasterxml.jackson.databind.node.ArrayNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import java.util.List;
 import org.springframework.stereotype.Component;
-import uk.gov.companieshouse.api.delta.ChildProperties;
 import uk.gov.companieshouse.api.delta.FilingHistory;
 import uk.gov.companieshouse.api.delta.FilingHistoryDelta;
 import uk.gov.companieshouse.api.filinghistory.InternalFilingHistoryApi;
 import uk.gov.companieshouse.filinghistory.consumer.mapper.posttransform.InternalFilingHistoryApiMapper;
 import uk.gov.companieshouse.filinghistory.consumer.mapper.posttransform.InternalFilingHistoryApiMapperArguments;
-import uk.gov.companieshouse.filinghistory.consumer.mapper.pretransform.ChildPair;
 import uk.gov.companieshouse.filinghistory.consumer.mapper.pretransform.PreTransformMapper;
 import uk.gov.companieshouse.filinghistory.consumer.transformrules.TransformerService;
 
 @Component
 public class FilingHistoryDeltaProcessor {
 
+    private static final List<String> childArrayKeys = List.of("annotations", "resolutions", "associated_filings");
     private final TransactionKindService kindService;
     private final PreTransformMapper preTransformMapper;
     private final TransformerService transformerService;
     private final InternalFilingHistoryApiMapper internalFilingHistoryApiMapper;
 
     public FilingHistoryDeltaProcessor(TransactionKindService kindService,
-                                       PreTransformMapper preTransformMapper,
-                                       TransformerService transformerService,
-                                       InternalFilingHistoryApiMapper internalFilingHistoryApiMapper) {
+            PreTransformMapper preTransformMapper,
+            TransformerService transformerService,
+            InternalFilingHistoryApiMapper internalFilingHistoryApiMapper) {
         this.kindService = kindService;
         this.preTransformMapper = preTransformMapper;
         this.transformerService = transformerService;
@@ -46,26 +44,25 @@ public class FilingHistoryDeltaProcessor {
 
         TransactionKindResult kindResult = kindService.encodeIdByTransactionKind(criteria);
 
-        ObjectNode topLevelObjectNode = preTransformMapper.mapDeltaToObjectNode(filingHistory);
-        ObjectNode transformedJsonNode = (ObjectNode) transformerService.transform(topLevelObjectNode, entityId);
+        ObjectNode preTransformNode = preTransformMapper.mapDeltaToObjectNode(kindResult.kind(), filingHistory);
 
-        if (!TOP_LEVEL.equals(kindResult.kind()) || hasChildArray(filingHistory.getChild())) {
-            ChildPair childPair = preTransformMapper.mapChildDeltaToObjectNode(kindResult.kind(), filingHistory);
-            ObjectNode dataNode = (ObjectNode) transformedJsonNode.get("data");
-            dataNode.putArray(childPair.childArrayKey()).add(transformerService.transform(childPair.node(), entityId));
-        }
+        ObjectNode transformedParentNode = (ObjectNode) transformerService.transform(preTransformNode, entityId);
+        ObjectNode dataNode = (ObjectNode) transformedParentNode.get("data");
+
+        childArrayKeys.forEach(key -> {
+            ArrayNode childArray = (ArrayNode) dataNode.get(key);
+            if (childArray != null && childArray.get(0) != null) {
+                dataNode.putArray(key).add(transformerService.transform(childArray.get(0), entityId));
+            }
+        });
 
         InternalFilingHistoryApiMapperArguments arguments = new InternalFilingHistoryApiMapperArguments(
-                transformedJsonNode,
+                transformedParentNode,
                 kindResult,
                 delta.getFilingHistory().getFirst().getCompanyNumber(),
                 delta.getDeltaAt(),
                 updatedBy);
 
         return internalFilingHistoryApiMapper.mapInternalFilingHistoryApi(arguments);
-    }
-
-    private static boolean hasChildArray(List<ChildProperties> childArray) {
-        return childArray != null && !childArray.isEmpty();
     }
 }
