+++ src/test/java/uk/gov/companieshouse/filinghistory/consumer/mapper/pretransform/ResolutionNodeMapperTest.java
@@ -6,34 +6,38 @@ import static org.mockito.Mockito.when;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ObjectNode;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import uk.gov.companieshouse.api.delta.DescriptionValues;
 import uk.gov.companieshouse.api.delta.FilingHistory;
+import uk.gov.companieshouse.filinghistory.consumer.transformrules.TransformerTestingUtils;
 import uk.gov.companieshouse.filinghistory.consumer.transformrules.functions.FormatDate;
 
 @ExtendWith(MockitoExtension.class)
 class ResolutionNodeMapperTest {
 
+    private static final ObjectMapper objectMapper = TransformerTestingUtils.getMapper();
     private static final String CASE_START_DATE = "01/01/2010";
     private static final String RES_TYPE = "res type";
     private static final String DESCRIPTION = "description";
     private static final String DATE = "02/02/2011";
     private static final String RESOLUTION_DATE = "03/03/2012";
-    @InjectMocks
-    private ResolutionNodeMapper resolutionNodeMapper;
 
-    @Mock
-    private ObjectMapper objectMapper;
+    private ResolutionNodeMapper resolutionNodeMapper;
     @Mock
     private FormatDate formatDate;
 
     @Mock
     private FilingHistory delta;
+
+    @BeforeEach
+    void setUp() {
+        resolutionNodeMapper = new ResolutionNodeMapper(objectMapper, formatDate);
+    }
 
     @ParameterizedTest
     @CsvSource({
@@ -41,8 +45,6 @@ class ResolutionNodeMapperTest {
             DESCRIPTION})
     void shouldMapResolutionObjectNode(final String description) {
         // given
-        ObjectNode actualObjectNode = new ObjectMapper().createObjectNode();
-        when(objectMapper.createObjectNode()).thenReturn(actualObjectNode);
         when(delta.getFormType()).thenReturn("form type");
         when(delta.getDescription()).thenReturn(description);
         when(formatDate.format(any())).thenReturn("date");
@@ -56,13 +58,19 @@ class ResolutionNodeMapperTest {
 
         when(delta.getDescriptionValues()).thenReturn(descriptionValues);
 
-        ObjectNode expectedObjectNode = new ObjectMapper()
-                .createObjectNode()
+        ObjectNode parentNode = objectMapper.createObjectNode();
+        parentNode.putObject("data");
+
+        ObjectNode expected = objectMapper.createObjectNode();
+        ObjectNode childNode = expected.putObject("data")
+                .putArray("resolutions")
+                .addObject();
+        childNode
                 .put("type", "form type")
                 .put("date", "date")
                 .put(DESCRIPTION, description);
 
-        expectedObjectNode
+        childNode
                 .putObject("description_values")
                 .put("case_start_date", CASE_START_DATE)
                 .put("res_type", RES_TYPE)
@@ -70,12 +78,10 @@ class ResolutionNodeMapperTest {
                 .put("date", DATE)
                 .put("resolution_date", RESOLUTION_DATE);
 
-        ChildPair expectedChildPair = new ChildPair("resolutions", expectedObjectNode);
-
         // when
-        ChildPair actualChildPair = resolutionNodeMapper.mapChildObjectNode(delta);
+        ObjectNode actual = resolutionNodeMapper.mapChildObjectNode(delta, parentNode);
 
         // then
-        assertEquals(expectedChildPair, actualChildPair);
+        assertEquals(expected, actual);
     }
 }
